67 - reorganize pages according to breadcrumbs & flow
