+++ pages/account-management.tsx
@@ -35,6 +35,7 @@ const AccountManagement: React.FunctionComponent = () => {
       })}
       data-h2-width="b(100) l(75)"
       data-h2-padding="b(all, none)"
+      crumbs={[{ title: "My passport" }, { title: "Manage your account" }]}
     >
       <Page>
         <LeftSection>
