+++ pages/manager/manager-dashboard.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { useIntl } from "react-intl";
-import Button, { colorMap } from "../../components/Button";
+import { colorMap } from "../../components/Button";
 import { PassportCard } from "../../components/PassportCard";
 import Layout from "../../components/Layout";
 import Page, { LeftSection, RightSection } from "../../components/Page";
@@ -9,16 +9,14 @@ import Link from "next/link";
 const ManagerDashboard: React.FunctionComponent = () => {
   const intl = useIntl();
 
-  const bold = (msg: string): React.ReactNode => (
-    <span data-h2-font-weight="b(700)">{msg}</span>
-  );
   return (
     <Layout
       title={intl.formatMessage({
         defaultMessage: "Manager dashboard",
       })}
       headTitle={intl.formatMessage({
-        defaultMessage: "Manager dashboard - GC Workplace Accessibility Passport",
+        defaultMessage:
+          "Manager dashboard - GC Workplace Accessibility Passport",
       })}
       crumbs={[{ title: "My Dashboard" }]}
     >
